requirements.txt
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
brotli-python==1.0.9
bzip2==1.0.8
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
deprecated==1.2.13
dill==0.3.8
Django==4.1
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
iniconfig==1.1.1
isort==5.13.2
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.0
optional-django==0.1.0
packaging==24.1
platformdirs==3.10.0
pluggy==1.0.0
pygments==2.15.1
pylint==3.2.7
pysocks==1.7.1
pytest==7.4.4
python-dotenv==0.21.0
python-tzdata==2023.3
redis-py==4.3.4
requests==2.32.3
rich==13.7.1
setuptools==72.1.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.4
tomlkit==0.13.2
tzdata==2024a
urllib3==2.2.2
wrapt==1.14.1